Create Hiding_text_in_images
This a simple program to hide text in an png images, it makes use of the least significant bit in the .png images ( which is generally considered as random noise and doesn't change with respect to the changes made to the image). The text is hidden in the image but the image looks 100% the same and human eye cannot find a difference in the image.